Moved Topbar State to another context, this fixes the broken plugin links
